move vis to cli
