+++ chronus/cli/__init__.py
@@ -6,7 +6,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from mpl_toolkits.mplot3d import Axes3D
 
-from chronus.model.Run import Run
+from chronus.domain.Run import Run
 
 
 def fake_data() -> list[Run]:
